+++ Mods/tech/HighDigitSAMs/Database/Vehicle/SAMPT_MC.lua
@@ -8,7 +8,9 @@ set_recursive_metatable(GT.chassis, GT_t.CH_t.STATIC);
 GT.visual.shape = "sampt_mc";
 GT.visual.shape_dstr = "sampt_standard_destroyed";
 
-GT.toggle_alarm_state_interval = 10;
+GT.IR_emission_coeff = 0.08
+
+GT.swing_on_run = false
 
 GT.visual.fire_size = 1.2; --relative burning size
 GT.visual.fire_pos[1] = 0; -- center of burn at long axis shift(meters)
@@ -22,20 +24,23 @@ GT.visual.min_time_agony       = 5
 GT.visual.max_time_agony       = 120
 GT.visual.agony_explosion_size = 5
 
+GT.visual.IR = { coeffs = {GT_t.IR_COEFFS_ENGINE_APC, GT_t.IR_COEFFS_WHEELS_Truck, {0.0, 0.0}, GT_t.IR_COEFFS_BODY, {0.1, 0.1/1000}}}
+
+GT.toggle_alarm_state_interval = 10.0;
+
 GT.sensor = {};
-GT.sensor.max_range_finding_target = 400000;
-GT.sensor.min_range_finding_target = 10;
-GT.sensor.max_alt_finding_target = 100000;
+GT.sensor.max_range_finding_target = 500000;
+GT.sensor.min_range_finding_target = 100;
+GT.sensor.max_alt_finding_target = 75000;
 GT.sensor.height = 5.895;
 
 GT.WS = {}
-GT.WS.maxTargetDetectionRange = 500;
-GT.WS.requiredUnits = {{"SAMPT_MGE", 1000, GT_t.REQUIRED_UNIT.NEED_AI_ON}};
+GT.WS.maxTargetDetectionRange = 500000;
 
 GT.Name = "SAMPT_MC";
-GT.DisplayName = _("SAMP/T Control Station");
-GT.DisplayNameShort = _("SAMP/T MC");
-GT.Rate = 20;
+GT.DisplayName = _("SAMP/T C2");
+GT.DisplayNameShort = _("SAMP/T C2");
+GT.Rate = 10;
 
 GT.EPLRS = true;
 
@@ -43,9 +48,10 @@ GT.DetectionRange = 0;
 GT.ThreatRange = 0;
 GT.mapclasskey = "P0091000046";
 GT.attribute = {wsType_Ground, wsType_SAM, wsType_NoWeapon, WSTYPE_PLACEHOLDER,
-"Trucks",
-"SAM CC",
-};
+                "Trucks",
+                "SAM CC",
+                };
+                
 GT.category = "Air Defence";
 GT.tags = {
     "Air Defence",