+++ Mods/tech/HighDigitSAMs/Database/Vehicle/SAMPT_ME.lua
@@ -8,7 +8,9 @@ set_recursive_metatable(GT.chassis, GT_t.CH_t.STATIC);
 GT.visual.shape = "sampt_me";
 GT.visual.shape_dstr = "sampt_standard_destroyed";
 
-GT.toggle_alarm_state_interval = 10;
+GT.toggle_alarm_state_interval = 10.0;
+GT.IR_emission_coeff = 0.08
+GT.visual.IR = { coeffs = {GT_t.IR_COEFFS_ENGINE_APC, GT_t.IR_COEFFS_WHEELS_Truck, {0.0, 0.0}, GT_t.IR_COEFFS_BODY, {0.1, 0.1/1000}}}
 
 GT.visual.fire_size = 1.2; --relative burning size
 GT.visual.fire_pos[1] = 0; -- center of burn at long axis shift(meters)
@@ -23,18 +25,16 @@ GT.visual.max_time_agony       = 120
 GT.visual.agony_explosion_size = 5
 
 GT.sensor = {};
-GT.sensor.max_range_finding_target = 400000;
-GT.sensor.min_range_finding_target = 10;
-GT.sensor.max_alt_finding_target = 100000;
+set_recursive_metatable(GT.sensor, GT_t.SN_visual);
 GT.sensor.height = 5.895;
 
 GT.WS = {}
-GT.WS.maxTargetDetectionRange = 500;
-GT.WS.requiredUnits = {{"SAMPT_MGE", 1000, GT_t.REQUIRED_UNIT.NEED_AI_ON}};
+GT.WS.maxTargetDetectionRange = 500000;
+GT.WS.requiredUnits = {{"SAMPT_MGE", 1000, 2}};
 
 GT.Name = "SAMPT_ME";
-GT.DisplayName = _("SAMP/T Engagement Station");
-GT.DisplayNameShort = _("SAMP/T ME");
+GT.DisplayName = _("SAMP/T ECS");
+GT.DisplayNameShort = _("SAMP/T ECS");
 GT.Rate = 20;
 
 GT.EPLRS = true;
@@ -43,9 +43,10 @@ GT.DetectionRange = 0;
 GT.ThreatRange = 0;
 GT.mapclasskey = "P0091000046";
 GT.attribute = {wsType_Ground, wsType_SAM, wsType_NoWeapon, WSTYPE_PLACEHOLDER,
-"Trucks",
-"SAM CC",
-};
+                "Trucks",
+                "SAM CC",
+                };
+
 GT.category = "Air Defence";
 GT.tags = {
     "Air Defence",