+++ Mods/tech/HighDigitSAMs/Liveries/SAMPT_MC/DESERT/description.lua
@@ -1,0 +1,6 @@
+livery = {
+	{"Material.079",	DIFFUSE			,	"Glass", true};
+	{"Material.155",	DIFFUSE			,	"ERO_SAMPT_DESERT", true};
+	{"Material.155",	DIFFUSE			,	"ERO_SAMPT_DESERT", true};
+}
+name = "Desert Camo"