+++ Mods/tech/HighDigitSAMs/Liveries/SAMPT_ME/DESERT/description.lua
@@ -1,0 +1,5 @@
+livery = {
+	{"Material.351",	DIFFUSE			,	"glass", true};
+	{"Material.157",	DIFFUSE			,	"ERO_SAMPT_DESERT", true};
+}
+name = "Desert Camo"