+++ Mods/tech/HighDigitSAMs/Liveries/SAMPT_ME/DIRT/description.lua
@@ -1,0 +1,5 @@
+livery = {
+	{"Material.351",	DIFFUSE			,	"glass", true};
+	{"Material.157",	DIFFUSE			,	"ERO_SAMPT_Earth", true};
+}
+name = "Dirt Camo"