+++ Mods/tech/HighDigitSAMs/Liveries/SAMPT_MGE/DIRT/description.lua
@@ -1,0 +1,6 @@
+livery = {
+	{"Material.388",	DIFFUSE			,	"ERO_SAMPT_Earth", true};
+	{"Material.079",	DIFFUSE			,	"Glass", true};
+	{"Material.157",	DIFFUSE			,	"ERO_SAMPT_Earth", true};
+}
+name = "Dirt Camo"