+++ Mods/tech/HighDigitSAMs/Liveries/SAMPT_MLT_Blk1/DIRT/description.lua
@@ -1,0 +1,16 @@
+livery = {
+	{"Material.024",	DIFFUSE			,	"ERO_SAMPT_Launch_Cannister", true};
+	{"Material.010",	DIFFUSE			,	"GLASS", true};
+	{"Material.024",	DIFFUSE			,	"ERO_SAMPT_Launch_Cannister", true};
+	{"Material.024",	DIFFUSE			,	"ERO_SAMPT_Launch_Cannister", true};
+	{"Material.084",	DIFFUSE			,	"ERO_SAMPT_Earth", true};
+	{"Material.083",	DIFFUSE			,	"ERO_SAMPT_Earth", true};
+	{"Material.024",	DIFFUSE			,	"ERO_SAMPT_Launch_Cannister", true};
+	{"Material.024",	DIFFUSE			,	"ERO_SAMPT_Launch_Cannister", true};
+	{"Material.024",	DIFFUSE			,	"ERO_SAMPT_Launch_Cannister", true};
+	{"Material.024",	DIFFUSE			,	"ERO_SAMPT_Launch_Cannister", true};
+	{"Material.024",	DIFFUSE			,	"ERO_SAMPT_Launch_Cannister", true};
+	{"Material.024",	DIFFUSE			,	"ERO_SAMPT_Launch_Cannister", true};
+	{"Material.024",	DIFFUSE			,	"ERO_SAMPT_Launch_Cannister", true};
+}
+name = "Dirt Camo"