+++ Mods/tech/HighDigitSAMs/Liveries/SAMPT_MRI_ARABEL/DIRT/description.lua
@@ -1,0 +1,7 @@
+livery = {
+	{"Material.011",	DIFFUSE			,	"ERO_SAMPT_Earth", true};
+	{"Material.082",	DIFFUSE			,	"ERO_SAMPT_Earth", true};
+	{"Material.049",	DIFFUSE			,	"grating", true};
+	{"Material.010",	DIFFUSE			,	"GLASS", true};
+}
+name = "Dirt Camo"