+++ Mods/tech/HighDigitSAMs/Liveries/SAMPT_MRI_GF300/DESERT/description.lua
@@ -1,0 +1,8 @@
+livery = {
+	{"Material.079",	DIFFUSE			,	"ERO_SAMPT_DESERT", true};
+	{"Material.306",	DIFFUSE			,	"ERO_SAMPT_DESERT", true};
+	{"Material.203",	DIFFUSE			,	"ERO_SAMPT_DESERT", true};
+	{"Material.010",	DIFFUSE			,	"GLASS", true};
+	{"Material.306",	DIFFUSE			,	"ERO_SAMPT_DESERT", true};
+}
+name = "Desert Camo"